services/processor/requirements.txt
kafka-python
psycopg2-binary
requests

scikit-learn
pandas
numpy
joblib